Fix the "bug" discovered by marcus in his app.t test
The async function is supposed to return a promise, which our initial
implementation did not do. I've rewritten the driver to do so and in
doing so I've had to rely more on Coro for scheduling than before.

This also required hooking into the Mojo::IOLoop to call `cede` on each
loop. This means that the loop is now essentially three phases (in some
order) check filehandles, check timers, check coroutines. I suspected
that we'd get here eventually.